// src/constants/Ui/colors/colors.types.ts
export type ColorsType = {
  // backgrounds
  background_primary: string;
  background_secondary: string;
  background_tertiary: string;
  background_Quaternary: string;
  background_Quinary: string;
  background_LightGreen: string;
  background_lightGray: string,

  // common
  primary: string;
  primaryDark: string;
  primaryLight: string;

  // card background
  card_bg_primary: string;

  // buttons
  // disabledPrimaryButton: string

  // texts
  contentPrimary: string;
  contentSecondary: string;
  contentTertiary: string;
  contentQuaternary: string;
  contentQuinary: string;
  contentGreen: string;
  contentDisabled: string;
  subtitle: string;
  primaryCtaText: string;
  textFade: string;
  secondaryCtaText: string;

  // borders
  border_1: string;
  border_active: string;
  border_2: string;
  // border_4: string
  // border_5: string

  // shadow color
  primaryShadow: string;

  // button
  disabledPrimaryButton: string;

  // icons
  primaryIconColor: string;
  secondaryIconColor: string;
  tertiaryIconColor: string;

  // searchbar
  searchbar: string;

  // primary-modal-backdrop-color
  // modalOverlay: string

  // map markers
  // mapPickupMarker: string,
  dropdown_primary: string;
};